# specs/002-professional-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Altavia Trade Professional Center API
  version: 1.0.0
paths:
...
  /api/profesional/productos/{product_id}:
    get: {summary: Ficha integral de producto, responses: {'200': {description: Ficha}, '404': {description: No encontrado}}}
  /api/profesional/clientes/{client_id}:
    get: {summary: Ficha integral de cliente, responses: {'200': {description: Ficha}, '404': {description: No encontrado}}}
...
  /api/profesional/metas:
    get: {summary: Listar metas, responses: {'200': {description: Metas}}}
    post: {summary: Crear meta, responses: {'201': {description: Meta creada}, '400': {description: Validación}}}
...
  /api/profesional/aprobaciones/{approval_id}/decision:
    post: {summary: Resolver aprobación, responses: {'200': {description: Resuelta}, '409': {description: Ya resuelta}}}
...
  /api/profesional/logistica/{request_id}:
    patch: {summary: Actualizar transportista, guía y entrega estimada, responses: {'200': {description: Seguimiento actualizado}, '400': {description: Datos inválidos}}}
...
  /api/profesional/lotes:
    get: {summary: Listar lotes y caducidad, responses: {'200': {description: Lotes}}}
    post: {summary: Registrar lote, responses: {'201': {description: Lote registrado}, '400': {description: Validación}}}
...
  /api/profesional/listas-precios:
    get: {summary: Listar precios por cliente o canal, responses: {'200': {description: Listas}}}
    post: {summary: Crear o actualizar lista de precios, responses: {'201': {description: Lista guardada}, '400': {description: Validación}}}
...